Add filesystem logging, bugfix oauth flow
OAuthlib has a bug where the callback URL protocol is insecure. Doing
a check of the protocol before fetching the token fixes the protocol
and prevents an error in production.

This can be avoided in a dev environment by setting the
OAUTH_INSECURE_TRANSPORT variable to 1, but it is not safe to use that
in prod.